review fixes wave 2: DIP, dedup, naming
#2: Extract generateMindId to standalone function. ExtensionLoader.cleanupExtensions
    as instance method. MindManager uses only type imports now.
#9: Shared resolveNodeModulesDir in sdkPaths.ts — used by sdkImport + factory.
#12: LoadResult→ExtensionLoadResult, toPublic→toExternalContext.
